fix offset in edge cases
This commit add the node shift in all appropriate places.
Mostly the shift when the parent is a vbox is added,
but this also has the annoying implication that we
need to ignore the page shift in LaTeX (added with
\moveright inside the shipout box), so \@outputpage is paged.
The main function is also a bit more efficient now.